delete & integer

+++ app/src/main/java/com/example/animalsoundrecognition/model/DataSoundParameters.kt
@@ -1,6 +1,6 @@
 package com.example.animalsoundrecognition.model
 
-class DataSoundParameters(val typeName:String, val signalEnvelope:List<Double>, val rootMeanSquareEnergy:List<Double>, val zeroCrossingDensity:Double) {
+class DataSoundParameters(val typeName:String, val signalEnvelope:List<Integer>, val rootMeanSquareEnergy:List<Integer>, val zeroCrossingDensity:Integer) {
     override fun toString(): String {
         return "Params(zeroCrossingDensity=$zeroCrossingDensity \n" +
                 "signalEnvelope=$signalEnvelope \n " +
